character fields privateStats and publicStats are now joined in stats
To ease access, statistics and status of characters are no longer available under different fields depending
if they are restricted or not. The two hooks, h/lotgd/crate-graphql/characterStats/private and /public still
exist and can be used accordingly for public information (like name, level, race) or private (like gold, gems).

Types extending BaseType can now call the method isGuarded to check whether or not access to the type is
restricted. This is used to only gather private stats if the user is actually authenticated to do so.